Switch the Java version to float (from double).

+++ src/com/nullprogram/noise/Noise.java
@@ -10,5 +10,5 @@ public interface Noise {
      * @param p  the position to query
      * @return the noise value at the selected position
      */
-    double sample(Vector p);
+    float sample(Vector p);
 }
